+++ tests/system_tests/test_hw_cpx_ap_i.py
@@ -598,6 +598,327 @@ def test_4iol_sdas_full(test_cpxap):
         param = m.read_fieldbus_parameters()
 
 
+def test_4iol_sdas_read_isdu_string_as_raw(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    ret = m.read_isdu(sdas_channel, 0x0010, 0)
+    # according to datasheet, this should return "Festo" as 64 byte string
+    # It actually returns "Festo AG & Co. KG" on my device, which could be a
+    # different hardware version.
+    # in io-link, strings are encoded byteorder "msb first", so no change required
+    ret_value = ret.decode("ascii")
+
+    # Assert
+    assert ret_value == "Festo AG & Co. KG"
+
+
+def test_4iol_sdas_read_isdu_string_as_str(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    ret = m.read_isdu(sdas_channel, 0x0010, 0, data_type="str")
+    # according to datasheet, this should return "Festo" as 64 byte string
+    # It actually returns "Festo AG & Co. KG" on my device, which could be a
+    # different hardware version.
+
+    # Assert
+    assert ret == "Festo AG & Co. KG"
+
+
+def test_4iol_sdas_write_isdu_string_as_raw(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    m.write_isdu(b"FESTO", sdas_channel, 0x0018, 0)
+    ret = m.read_isdu(sdas_channel, 0x0018, 0)
+
+    # This splits at the first b"\x00" and only uses the left side.
+    ret_value = ret.decode("ascii").split("\x00", 1)[0]
+
+    # Assert
+    assert ret_value == "FESTO"
+
+    # Act more
+    # This should test if there are characters remaining in the isdu
+    # from the "FESTO" and if they are returned with it incorrectly
+    m.write_isdu(b"LOL", sdas_channel, 0x0018, 0)
+    ret = m.read_isdu(sdas_channel, 0x0018, 0)
+
+    # This splits at the first b"\x00" and only uses the left side.
+    ret_value = ret.decode("ascii").split("\x00", 1)[0]
+
+    # Assert
+    assert ret_value == "LOL"
+
+
+def test_4iol_sdas_write_isdu_string_as_str(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    m.write_isdu("FESTO", sdas_channel, 0x0018, 0)
+    ret = m.read_isdu(sdas_channel, 0x0018, 0, data_type="str")
+
+    # Assert
+    assert ret == "FESTO"
+
+    # Act more
+    # This should test if there are characters remaining in the isdu
+    # from the "FESTO" and if they are returned with it incorrectly
+    m.write_isdu("LOL", sdas_channel, 0x0018, 0)
+    ret = m.read_isdu(sdas_channel, 0x0018, 0, data_type="str")
+
+    # Assert
+    assert ret == "LOL"
+
+
+def test_4iol_sdas_write_isdu_string_1byte(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    m.write_isdu("FESTO", sdas_channel, 0x0018, 0)
+    ret = m.read_isdu(sdas_channel, 0x0018, 0, data_type="str")
+
+    # Assert
+    assert ret == "FESTO"
+
+    # Act more
+    # This should test if there are characters remaining in the isdu
+    # from the "FESTO" and if they are returned with it incorrectly
+    m.write_isdu("x", sdas_channel, 0x0018, 0)
+    ret = m.read_isdu(sdas_channel, 0x0018, 0, data_type="str")
+
+    # Assert
+    assert ret == "x"
+
+
+def test_4iol_sdas_readwrite_isdu_int16_as_raw(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    m.write_isdu(b"\x01\x23", sdas_channel, 0x003C, 1)
+    ret = m.read_isdu(sdas_channel, 0x003C, 1)
+
+    ret_value = int.from_bytes(ret, byteorder="big")
+
+    # Assert
+    assert ret_value == 0x0123
+
+
+def test_4iol_sdas_readwrite_isdu_int16_as_int(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    m.write_isdu(0x0123, sdas_channel, 0x003C, 1)
+    ret = m.read_isdu(sdas_channel, 0x003C, 1, data_type="int")
+
+    # Assert
+    assert ret == 0x0123
+
+
+def test_4iol_sdas_readwrite_isdu_int8_as_raw(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    m.write_isdu(b"\x01", sdas_channel, 0x003D, 1)
+    ret = m.read_isdu(sdas_channel, 0x003D, 1)
+
+    ret_value = int.from_bytes(ret, byteorder="big")
+
+    # Assert
+    assert ret_value == 0x01
+
+
+def test_4iol_sdas_readwrite_isdu_int8_as_int(test_cpxap):
+    m = test_cpxap.modules[4]
+    sdas_channel = 0
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[sdas_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act
+    m.write_isdu(1, sdas_channel, 0x003D, 1)
+    ret = m.read_isdu(sdas_channel, 0x003D, 1, data_type="int")
+
+    # Assert
+    assert ret == 1
+
+
+def test_4iol_emcs_read_isdu_str(test_cpxap):
+    m = test_cpxap.modules[4]
+    emcs_channel = 3
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", emcs_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[emcs_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act & Assert
+    assert m.read_isdu(emcs_channel, 16, data_type="str") == "Festo"
+
+
+def test_4iol_emcs_read_isdu_bool(test_cpxap):
+    m = test_cpxap.modules[4]
+    emcs_channel = 3
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", emcs_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[emcs_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act & Assert
+    assert m.read_isdu(emcs_channel, 259, data_type="bool") is False
+
+
+def test_4iol_emcs_write_isdu_bool(test_cpxap):
+    m = test_cpxap.modules[4]
+    emcs_channel = 3
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", emcs_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[emcs_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act & Assert
+    m.write_isdu(True, emcs_channel, 259)
+    assert m.read_isdu(emcs_channel, 259, data_type="bool") is True
+
+    # Reset
+    m.write_isdu(False, emcs_channel, 259)
+    assert m.read_isdu(emcs_channel, 259, data_type="bool") is False
+
+
+def test_4iol_emcs_read_int32_with_move(test_cpxap):
+    m = test_cpxap.modules[4]
+    emcs_channel = 3
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", emcs_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[emcs_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # Act & Assert
+    m.write_channel(emcs_channel, b"\x00\x02")  # Move
+    # wait for move to finish
+    while not int.from_bytes(m.read_channel(emcs_channel), byteorder="big") & 0x02:
+        time.sleep(0.01)
+
+    for _ in range(10):  # wait some more
+        time.sleep(0.05)
+        m.read_channel(emcs_channel)
+
+    assert m.read_isdu(emcs_channel, 288, data_type="int") > 0x00FFFFFF
+
+    m.write_channel(emcs_channel, b"\x00\x01")  # Move
+    # wait for move to finish
+    while not int.from_bytes(m.read_channel(emcs_channel), byteorder="big") & 0x01:
+        time.sleep(0.01)
+
+    for _ in range(10):  # wait some more
+        time.sleep(0.05)
+        m.read_channel(emcs_channel)
+
+    assert m.read_isdu(emcs_channel, 288, data_type="int") < 0xFF
+
+
+def test_4iol_emcs_read_isdu_float(test_cpxap):
+    m = test_cpxap.modules[4]
+    emcs_channel = 3
+
+    # Setup
+    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", emcs_channel)
+    time.sleep(0.05)
+    param = m.read_fieldbus_parameters()
+    while param[emcs_channel]["Port status information"] != "OPERATE":
+        param = m.read_fieldbus_parameters()
+
+    # read intermediate position
+    ret = m.read_isdu(emcs_channel, 264, data_type="float") * 0.01
+    assert isinstance(ret, float)
+    assert 0.01 < ret < 0.03
+
+    # read end postion out
+    ret = m.read_isdu(emcs_channel, 262, data_type="float") * 0.01
+    assert isinstance(ret, float)
+    assert 0.01 < ret < 0.03
+
+
 def test_4iol_ehps(test_cpxap):
     m = test_cpxap.modules[4]
 
@@ -760,7 +1081,7 @@ def test_4iol_ethrottle_isdu_read(test_cpxap):
     while param[ethrottle_channel]["Port status information"] != "OPERATE":
         param = m.read_fieldbus_parameters()
 
-    assert (m.read_isdu(ethrottle_channel, 16, 0)[:5]) == b"Festo"
+    assert (m.read_isdu(ethrottle_channel, 16, 0, data_type="str")) == "Festo"
 
     m.write_module_parameter("Port Mode", "DEACTIVATED", ethrottle_channel)
     # wait for inactive
@@ -783,12 +1104,16 @@ def test_4iol_ethrottle_isdu_write_1byte(test_cpxap):
         param = m.read_fieldbus_parameters()
 
     function_tag_idx = 25
-    # write something different first
-    m.write_isdu(b"\x12\x34", ethrottle_channel, function_tag_idx, 0)
+    # write zeroes first. This is because this prototype doesn't handle the lenght
+    # register correctly for the parameters. Otherwise only relevant bytes would be
+    # transferred, see sdas isdu write/read testst
+    m.write_isdu(b"\x00" * 32, ethrottle_channel, function_tag_idx, 0)
     time.sleep(0.05)
     m.write_isdu(b"\xCA", ethrottle_channel, function_tag_idx, 0)
 
-    assert m.read_isdu(ethrottle_channel, function_tag_idx, 0)[:2] == b"\xCA\x00"
+    assert m.read_isdu(ethrottle_channel, function_tag_idx, 0) == b"\xCA" + b"\x00" * 31
+    # we can also cut the rest off in this case
+    assert m.read_isdu(ethrottle_channel, function_tag_idx, 0)[:1] == b"\xCA"
 
     m.write_module_parameter("Port Mode", "DEACTIVATED", ethrottle_channel)
     # wait for inactive