+++ tests/unit_tests/cpx_system/cpx_ap/test_ap_module.py
@@ -1935,7 +1935,7 @@ class TestApModule:
         ]
 
     @pytest.mark.parametrize("input_value", [0, 1, 2, 3])
-    def test_read_isdu(self, module_fixture, input_value):
+    def test_read_isdu_different_channels(self, module_fixture, input_value):
         """Test read_isdu"""
         # Arrange
         module = module_fixture
@@ -1963,9 +1963,17 @@ class TestApModule:
                 call(b"\x64\x00", 34001),  # COMMAND (read 100)
             ]
         )
-        module.base.read_reg_data.assert_has_calls([call(34000, 1), call(34007, 119)])
+        module.base.read_reg_data.assert_has_calls(
+            [
+                call(34000, 1),
+                call(34006),
+                call(34007, 0),
+            ]
+        )
 
-        assert result == b"\x00\x00"
+        assert (
+            result == b""
+        )  # will cut to the actual_length which is returned with 0 in this test
 
     def test_read_isdu_no_response(self, module_fixture):
         """Test read_isdu"""
@@ -2044,17 +2052,16 @@ class TestApModule:
         subindex = 5  # random
 
         module.write_isdu(data, channel, index, subindex)
+        length = len(data)
 
         # Assert
-        length_to_write = len(data) if not len(data) % 2 else len(data) + 1
-
         module.base.write_reg_data.assert_has_calls(
             [
                 call(b"\x01\x00", 34002),  # MODULE_NO (position add 1)
                 call((channel + 1).to_bytes(2, "little"), 34003),  # CHANNEL (add 1)
                 call(b"\x04\x00", 34004),  # INDEX
                 call(b"\x05\x00", 34005),  # SUBINDEX
-                call(length_to_write.to_bytes(2, "little"), 34006),  # LENGTH
+                call(length.to_bytes(2, "little"), 34006),  # LENGTH
                 call(data, 34007),  # DATA
                 call(b"\x65\x00", 34001),  # COMMAND (read 101)
             ]
@@ -2079,3 +2086,68 @@ class TestApModule:
 
         with pytest.raises(CpxRequestError):
             module.write_isdu(data, channel, index, subindex)
+
+    @pytest.mark.parametrize(
+        "input_value, expected_output",
+        [("str", ""), ("int", 0), ("raw", b""), ("bool", False)],
+    )
+    def test_read_isdu_different_datatypes(
+        self, module_fixture, input_value, expected_output
+    ):
+        """Test read_isdu"""
+        # Arrange
+        module = module_fixture
+        module.position = 0
+        module.apdd_information.product_category = ProductCategory.IO_LINK.value
+        module.base = Mock()
+        module.base.write_reg_data = Mock()
+        module.base.read_reg_data = Mock(return_value=b"\x00\x00")
+
+        # Act
+        ret = module.read_isdu(0, 0, data_type=input_value)
+
+        # Assert
+        assert ret == expected_output
+
+    @pytest.mark.parametrize(
+        "input_value,, length, expected_output",
+        [
+            ("str", 3, b"str"), # string
+            (1, 1, b"\x01"), # int8
+            (0xCAFE, 2, b"\xFE\xCA"), # int16
+            (0xBEBAFECA, 4 , b"\xCA\xFE\xBA\xBE"), # int32
+            (b"\xCA\xFE", 2, b"\xCA\xFE"), # bytes = raw
+            (True, 1, b"\x01"), # bool true
+            (False, 1, b"\x00"), # bool false
+        ],
+    )
+    def test_write_isdu_different_datatypes(
+        self, module_fixture, input_value, length, expected_output
+    ):
+        """Test read_isdu"""
+        # Arrange
+        module = module_fixture
+        module.position = 0
+        module.apdd_information.product_category = ProductCategory.IO_LINK.value
+        module.base = Mock()
+        module.base.write_reg_data = Mock()
+        module.base.read_reg_data = Mock(return_value=b"\x00\x00")
+
+        # Act
+        module.write_isdu(input_value, 0, 0)
+
+        command = 51 if isinstance(input_value, (bool, int)) else 101
+
+        # Assert
+        module.base.write_reg_data.assert_has_calls(
+            [
+                call(b"\x01\x00", 34002),  # MODULE_NO (position add 1)
+                call((1).to_bytes(2, "little"), 34003),  # CHANNEL (add 1)
+                call(b"\x00\x00", 34004),  # INDEX
+                call(b"\x00\x00", 34005),  # SUBINDEX
+                call(length.to_bytes(2, "little"), 34006),  # LENGTH
+                call(expected_output, 34007),  # DATA
+                call(command.to_bytes(2, "little"), 34001),  # COMMAND
+            ]
+        )
+        module.base.read_reg_data.assert_has_calls([call(34000, 1)])